Move Oleville to the news section
Oleville view was really news, and I like having an even number of front-page buttons.
