+++ app.js
@@ -35,8 +35,8 @@ import NewsItemView from './views/news/news-item'
 import SISView from './views/sis'
 import {BuildingHoursView, BuildingHoursDetailView} from './views/building-hours'
 import TransportationView from './views/transportation'
-import OlevilleView from './views/oleville'
-import OlevilleNewsStoryView from './views/oleville/latestView'
+import OlevilleView from './views/news/oleville-view'
+import OlevilleNewsStoryView from './views/news/latestView'
 import SettingsView from './views/settings'
 import SISLoginView from './views/settings/login'
 import CreditsView from './views/settings/credits'
