+++ views/news/latestView.js
