--- views/oleville/types.js
@@ -1,8 +1,0 @@
-import {Navigator} from 'react-native'
-
-export type OlevilleLatestPropsType = {
-  navigator: typeof Navigator,
-  title: string,
-  imageURL: string,
-  content: string,
-};
